move part of dashboard-details to a new component dedicated to the display of sections
