+++ internal/api/front/perses/src/app/project/dashboard/dashboard-details/dashboard-sections/dashboard-sections.component.html
@@ -1,0 +1,21 @@
+<ng-container *ngFor="let section of spec.sections">
+  <h2>{{section.name}}</h2>
+  <div *ngFor="let panel of section.panels">
+    <h3>{{panel.name}}</h3>
+    <div>
+      <ngx-charts-line-chart
+      [view]="view"
+      [scheme]="colorScheme"
+      [results]="getChartData(section.name, panel.name)"
+      [gradient]="gradient"
+      [xAxis]="showXAxis"
+      [yAxis]="showYAxis"
+      [legend]="showLegend"
+      [showXAxisLabel]="showXAxisLabel"
+      [showYAxisLabel]="showYAxisLabel"
+      [xAxisLabel]="xAxisLabel"
+      [yAxisLabel]="yAxisLabel">
+      </ngx-charts-line-chart>
+    </div>
+  </div>
+</ng-container>
